requirements.txt
fastapi
uvicorn[standard]
openai
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
pydantic
pydantic-settings
python-multipart
sqlalchemy
psycopg2-binary
alembic
